Weaken mypy to ignore dancing_links structures
Knuth's Dancing Links is technically not type-safe as a cycle of links
may contain Root and Column (horizontal), or Column and DataObject (vertical).
We will avoid type hinting these too much as it adds clutter for not much gain,
as the correctness is guaranteed by the algorithm and its usage.